refactor(browser): ADR-023 Phase 1 trunk — extract resolver candidate-collection (bit-equal)
S1: new src/tools/browser-resolver.ts exports buildCandidateCollectionJs — a VERBATIM extraction of browser_search's inline candidate-collection IIFE (the shared resolver core later phases layer actionability / climb / coords onto for by-axis browser_click/fill). S7: browser_search now delegates to it; generated JS is byte-equal so the public contract is unchanged (NFR-1 / AC-9). Pinned by tests/unit/browser-resolver-candidate-collection.test.ts (full-template snapshot + interpolation + scoring-constant guards, R-P1-1 snapshot-first). browser.ts -181 lines, no behavior change.

Plan: desktop-touch-mcp-internal@aa6fcd3:docs/adr-023-phase-1-resolver-plan.md S1/S7.
