+++ src/tools/browser-resolver.ts
@@ -1,0 +1,228 @@
+/**
+ * browser-resolver.ts — ADR-023 Phase 1 resolver core.
+ *
+ * The shared injected-JS builder for semantic element resolution. Phase 1 trunk
+ * (S1/S7) extracts the candidate-collection IIFE VERBATIM from browser_search so
+ * `browser_search` stays bit-equal (snapshot-pinned by
+ * tests/unit/browser-resolver-candidate-collection.test.ts) while later phases
+ * layer actionability / ancestor climb / physical-coord resolution on top for
+ * browser_click({by}) / browser_fill({by}).
+ *
+ * The IIFE returns `{ total, returned, truncated, results[] }` (each result:
+ * type / text / selector / role / ariaLabel / matchedBy / confidence /
+ * inViewport / rect) or `{ __error }` (ScopeNotFound / InvalidRegex / Timeout).
+ *
+ * Plan: desktop-touch-mcp-internal:docs/adr-023-phase-1-resolver-plan.md S1/S7.
+ */
+
+export interface CandidateCollectionArgs {
+  by: "text" | "regex" | "role" | "ariaLabel" | "selector";
+  pattern: string;
+  scope?: string;
+  maxResults: number;
+  offset: number;
+  visibleOnly: boolean;
+  inViewportOnly: boolean;
+  caseSensitive: boolean;
+}
+
+/**
+ * Build the injected-JS IIFE that collects, scores, filters and shapes candidate
+ * elements for `browser_search`. This is a VERBATIM extraction of the former
+ * inline template in `browserSearchHandler` — the generated string is byte-equal
+ * (pinned by snapshot) so the public `browser_search` contract is unchanged
+ * (NFR-1 / AC-9). Do not change the emitted JS without updating the snapshot.
+ */
+export function buildCandidateCollectionJs(args: CandidateCollectionArgs): string {
+  const { by, pattern, scope, maxResults, offset, visibleOnly, inViewportOnly, caseSensitive } = args;
+  return `
+(function() {
+  const root = ${scope ? `document.querySelector(${JSON.stringify(scope)})` : "document"};
+  if (!root) return { __error: "ScopeNotFound" };
+
+  const by = ${JSON.stringify(by)};
+  const pat = ${JSON.stringify(pattern)};
+  const cs  = ${JSON.stringify(caseSensitive)};
+  const visibleOnly = ${JSON.stringify(visibleOnly)};
+  const viewportOnly = ${JSON.stringify(inViewportOnly)};
+  const maxN = ${JSON.stringify(maxResults + offset)};
+  const offN = ${JSON.stringify(offset)};
+
+  function isVisible(el) {
+    const s = window.getComputedStyle(el);
+    if (s.display === 'none' || s.visibility === 'hidden' || s.opacity === '0') return false;
+    const r = el.getBoundingClientRect();
+    return r.width > 0 && r.height > 0;
+  }
+  function inViewportRect(rect) {
+    return rect.top < window.innerHeight && rect.bottom > 0 &&
+           rect.left < window.innerWidth && rect.right > 0;
+  }
+  function bestSelector(el) {
+    if (el.id) return '#' + CSS.escape(el.id);
+    const name = el.getAttribute('name');
+    if (name) return el.tagName.toLowerCase() + '[name=' + JSON.stringify(name) + ']';
+    const aria = el.getAttribute('aria-label');
+    if (aria && aria.length < 80)
+      return el.tagName.toLowerCase() + '[aria-label=' + JSON.stringify(aria) + ']';
+    for (const attr of ['data-testid', 'data-asin']) {
+      const v = el.getAttribute(attr);
+      if (v && v.length < 60) return el.tagName.toLowerCase() + '[' + attr + '=' + JSON.stringify(v) + ']';
+    }
+    let node = el; let path = '';
+    for (let depth = 0; depth < 2 && node.parentElement; depth++) {
+      const p = node.parentElement;
+      const idx = Array.from(p.children).indexOf(node) + 1;
+      const seg = node.tagName.toLowerCase() + ':nth-child(' + idx + ')';
+      path = path ? seg + ' > ' + path : seg;
+      if (p.id) { path = '#' + CSS.escape(p.id) + ' > ' + path; break; }
+      node = p;
+    }
+    return path || el.tagName.toLowerCase();
+  }
+  function classify(el) {
+    const tag = el.tagName.toLowerCase();
+    if (tag === 'a') return 'link';
+    if (tag === 'button' || el.getAttribute('role') === 'button') return 'button';
+    if (tag === 'input' || tag === 'textarea' || tag === 'select') return 'input';
+    if (/^h[1-6]$/.test(tag)) return 'heading';
+    if (tag === 'p' || tag === 'span' || tag === 'div') return 'text';
+    return 'other';
+  }
+  function elText(el) {
+    const t = (el.textContent || '').trim().replace(/\\s+/g, ' ').slice(0, 80);
+    if (!t && el.tagName === 'INPUT')
+      return (el.placeholder || el.value || el.getAttribute('aria-label') || '').slice(0, 80);
+    return t;
+  }
+  function score(matched, visible) {
+    let s = matched;
+    if (!visible) s = Math.max(0, s - 0.3);
+    return Math.round(s * 100) / 100;
+  }
+
+  // Bound the scan — pages can have 10k+ nodes and CDP timeout is 15s.
+  const SCAN_BUDGET_MS = 3000;
+  const nowFn = (typeof performance !== 'undefined' ? () => performance.now() : () => Date.now());
+  const startTs = nowFn();
+  const deadline = startTs + SCAN_BUDGET_MS;
+  let aborted = false;
+  // Sample the clock every 1024 iterations — cheap but keeps latency bounded.
+  function overBudget(i) { return (i & 0x3FF) === 0 && nowFn() > deadline; }
+
+  // IIFE-local match-state stores. WeakMap is essential: DOM elements persist
+  // across Runtime.evaluate calls, so any expando we set (e.g. el.__matchScore)
+  // would leak into the next search and contaminate scores / matchedBy / dedupe.
+  // WeakMap is GC'd at IIFE end so each call starts clean.
+  const matchScore = new WeakMap();
+  const matchedByMap = new WeakMap();
+  const pushed = new Set();
+  function record(el, score, by) {
+    const prev = matchScore.get(el) || 0;
+    if (score > prev) { matchScore.set(el, score); matchedByMap.set(el, by); }
+    if (!pushed.has(el)) { candidates.push(el); pushed.add(el); }
+  }
+
+  const all = root.querySelectorAll('*');
+  let candidates = [];
+
+  if (by === 'selector') {
+    const selectorMatches = Array.from(root.querySelectorAll(pat));
+    for (let i = 0; i < selectorMatches.length; i++) {
+      if (overBudget(i)) { aborted = true; break; }
+      record(selectorMatches[i], 1.0, 'selector');
+    }
+  } else if (by === 'text') {
+    const needle = cs ? pat : pat.toLowerCase();
+    let i = 0;
+    for (const el of all) {
+      if (overBudget(i++)) { aborted = true; break; }
+      // Direct child text only (avoid double-counting parent matches via descendants)
+      const direct = Array.from(el.childNodes)
+        .filter(n => n.nodeType === 3)
+        .map(n => n.textContent || '')
+        .join('').trim();
+      if (!direct) continue;
+      const hay = cs ? direct : direct.toLowerCase();
+      if (hay === needle) record(el, 1.0, 'text');
+      else if (hay.includes(needle)) record(el, 0.8, 'text');
+    }
+  } else if (by === 'regex') {
+    let re;
+    try { re = new RegExp(pat, (cs ? '' : 'i') + 'u'); }
+    catch (e) { return { __error: "InvalidRegex", message: String(e) }; }
+    let i = 0;
+    for (const el of all) {
+      if (overBudget(i++)) { aborted = true; break; }
+      const direct = Array.from(el.childNodes).filter(n => n.nodeType === 3).map(n => n.textContent || '').join('').trim();
+      if (!direct) continue;
+      if (re.test(direct)) record(el, 0.9, 'regex');
+    }
+  } else if (by === 'role') {
+    const needle = cs ? pat : pat.toLowerCase();
+    let i = 0;
+    for (const el of all) {
+      if (overBudget(i++)) { aborted = true; break; }
+      const role = el.getAttribute('role') || '';
+      const cmp = cs ? role : role.toLowerCase();
+      if (cmp === needle) record(el, 0.75, 'role');
+    }
+    // Implicit roles — score slightly higher because they're guaranteed by tag.
+    if (!aborted && needle === 'button')  for (const el of root.querySelectorAll('button')) record(el, 0.85, 'roleImplicit');
+    if (!aborted && needle === 'link')    for (const el of root.querySelectorAll('a[href]')) record(el, 0.85, 'roleImplicit');
+    if (!aborted && needle === 'heading') for (const el of root.querySelectorAll('h1,h2,h3,h4,h5,h6')) record(el, 0.85, 'roleImplicit');
+  } else if (by === 'ariaLabel') {
+    const needle = cs ? pat : pat.toLowerCase();
+    let i = 0;
+    for (const el of all) {
+      if (overBudget(i++)) { aborted = true; break; }
+      const aria = el.getAttribute('aria-label') || '';
+      if (!aria) continue;
+      const cmp = cs ? aria : aria.toLowerCase();
+      if (cmp === needle) record(el, 0.95, 'ariaLabel');
+      else if (cmp.includes(needle)) record(el, 0.7, 'ariaLabel');
+    }
+  }
+
+  // candidates already de-duplicated via the pushed Set in record()
+
+  if (aborted && candidates.length === 0) {
+    return { __error: "Timeout", message: "Scan budget exceeded with no matches; narrow scope or maxResults." };
+  }
+
+  const filtered = [];
+  for (const el of candidates) {
+    const visible = isVisible(el);
+    if (visibleOnly && !visible) continue;
+    const rect = el.getBoundingClientRect();
+    const inVp = inViewportRect(rect);
+    if (viewportOnly && !inVp) continue;
+    filtered.push({ el, visible, rect, inVp });
+  }
+
+  // Score and sort by confidence desc
+  filtered.sort((a, b) => {
+    const sa = score(matchScore.get(a.el) || 0, a.visible);
+    const sb = score(matchScore.get(b.el) || 0, b.visible);
+    return sb - sa;
+  });
+
+  const total = filtered.length;
+  const sliced = filtered.slice(offN, offN + (maxN - offN));
+
+  const results = sliced.map(({ el, visible, rect, inVp }) => ({
+    type: classify(el),
+    text: elText(el),
+    selector: bestSelector(el),
+    role: el.getAttribute('role') || undefined,
+    ariaLabel: el.getAttribute('aria-label') || undefined,
+    matchedBy: matchedByMap.get(el),
+    confidence: score(matchScore.get(el) || 0, visible),
+    inViewport: inVp,
+    rect: { x: Math.round(rect.left), y: Math.round(rect.top), w: Math.round(rect.width), h: Math.round(rect.height) },
+  }));
+
+  return { total, returned: results.length, truncated: total > offN + results.length, results };
+})()
+`;
+}
